Fix worsened variable name
In the previous commit https://github.com/iml130/firos/commit/9eeb43644a4f7136ee80251f5f2e4251cb20921e the adjustment of a variable name led
to an unexpected behaviour. This is caused by a conflict between
two variables of the same name 'folder'. This consequently led to
the fact that all puplisher and subscriber could not be imported.